backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
python-multipart==0.0.20
pillow==11.3.0
numpy==2.3.2
pydantic-settings==2.10.1
replicate==1.0.7
pytest==8.4.1
